DS: adjust memory control
