+++ utils/dynstream/memory_control.go
@@ -47,6 +47,7 @@ type areaMemStat[A Area, P Path, T Event, D Dest, H Handler[A, P, T, D]] struct 
 	totalPendingSize     atomic.Int64
 	paused               atomic.Bool
 	lastSendFeedbackTime atomic.Value
+	algorithm            MemoryControlAlgorithm
 }
 
 func newAreaMemStat[A Area, P Path, T Event, D Dest, H Handler[A, P, T, D]](
@@ -61,6 +62,7 @@ func newAreaMemStat[A Area, P Path, T Event, D Dest, H Handler[A, P, T, D]](
 		memControl:           memoryControl,
 		feedbackChan:         feedbackChan,
 		lastSendFeedbackTime: atomic.Value{},
+		algorithm:            NewMemoryControlAlgorithm(settings.algorithm),
 	}
 	res.lastSendFeedbackTime.Store(time.Unix(0, 0))
 	res.settings.Store(&settings)
@@ -121,25 +123,13 @@ func (as *areaMemStat[A, P, T, D, H]) memoryUsageRatio() float64 {
 // It needs to be called after a event is appended.
 // Note: Our gaol is to fast pause, and lazy resume.
 func (as *areaMemStat[A, P, T, D, H]) updatePathPauseState(path *pathInfo[A, P, T, D, H]) {
-	algorithm := as.settings.Load().algorithm
-	var pause, resume bool
-	var memoryUsageRatio float64
-	switch algorithm {
-	case MemoryControlForEventCollector:
-		pause, resume, memoryUsageRatio = shouldPausePathV2(
-			path.paused.Load(),
-			path.pendingSize.Load(),
-			as.totalPendingSize.Load(),
-			as.settings.Load().maxPendingSize,
-			as.pathCount.Load(),
-		)
-	default:
-		pause, resume, memoryUsageRatio = shouldPausePath(
-			path.paused.Load(),
-			path.pendingSize.Load(),
-			as.settings.Load().maxPendingSize,
-		)
-	}
+	pause, resume, memoryUsageRatio := as.algorithm.ShouldPausePath(
+		path.paused.Load(),
+		path.pendingSize.Load(),
+		as.totalPendingSize.Load(),
+		as.settings.Load().maxPendingSize,
+		as.pathCount.Load(),
+	)
 
 	sendFeedback := func(pause bool) {
 		now := time.Now()
@@ -186,23 +176,11 @@ func (as *areaMemStat[A, P, T, D, H]) updatePathPauseState(path *pathInfo[A, P, 
 }
 
 func (as *areaMemStat[A, P, T, D, H]) updateAreaPauseState(path *pathInfo[A, P, T, D, H]) {
-	algorithm := as.settings.Load().algorithm
-	var pause, resume bool
-	var memoryUsageRatio float64
-	switch algorithm {
-	case MemoryControlForEventCollector:
-		pause, resume, memoryUsageRatio = shouldPauseAreaV2(
-			as.paused.Load(),
-			as.totalPendingSize.Load(),
-			as.settings.Load().maxPendingSize,
-		)
-	default:
-		pause, resume, memoryUsageRatio = shouldPauseArea(
-			as.paused.Load(),
-			as.totalPendingSize.Load(),
-			as.settings.Load().maxPendingSize,
-		)
-	}
+	pause, resume, memoryUsageRatio := as.algorithm.ShouldPauseArea(
+		as.paused.Load(),
+		as.totalPendingSize.Load(),
+		as.settings.Load().maxPendingSize,
+	)
 
 	sendFeedback := func(pause bool) {
 		now := time.Now()
@@ -236,7 +214,7 @@ func (as *areaMemStat[A, P, T, D, H]) updateAreaPauseState(path *pathInfo[A, P, 
 		)
 	}
 
-	if algorithm != MemoryControlForEventCollector {
+	if as.settings.Load().algorithm != MemoryControlForEventCollector {
 		failpoint.Inject("PauseArea", func() {
 			log.Warn("inject PauseArea")
 			sendFeedback(true)
@@ -250,7 +228,7 @@ func (as *areaMemStat[A, P, T, D, H]) updateAreaPauseState(path *pathInfo[A, P, 
 		sendFeedback(false)
 	}
 
-	if algorithm == MemoryControlForEventCollector && as.paused.Load() {
+	if as.settings.Load().algorithm == MemoryControlForEventCollector && as.paused.Load() {
 		log.Panic("area is paused, but the algorithm is v2, this should not happen", zap.String("component", as.settings.Load().component))
 	}
 }
@@ -360,150 +338,4 @@ func (a *AreaMemoryMetric[A]) MaxMemory() int64 {
 
 func (a *AreaMemoryMetric[A]) Area() A {
 	return a.area
-}
-
-// shouldPausePath determines if a path should be paused based on memory usage.
-// If the memory usage is greater than the 20% of max pending size, the path should be paused.
-func shouldPausePath(
-	paused bool,
-	pendingSize int64,
-	maxPendingSize uint64,
-) (pause bool, resume bool, memoryUsageRatio float64) {
-	memoryUsageRatio = float64(pendingSize) / float64(maxPendingSize)
-
-	switch {
-	case paused:
-		// If the path is paused, we only need to resume it when the memory usage is less than 10%.
-		if memoryUsageRatio < 0.1 {
-			resume = true
-		}
-	default:
-		// If the path is not paused, we need to pause it when the memory usage is greater than 30% of max pending size.
-		if memoryUsageRatio >= 0.2 {
-			pause = true
-		}
-	}
-
-	return pause, resume, memoryUsageRatio
-}
-
-// shouldPauseArea determines if the area should be paused based on memory usage.
-// If the memory usage is greater than the 80% of max pending size, the area should be paused.
-func shouldPauseArea(
-	paused bool,
-	pendingSize int64,
-	maxPendingSize uint64,
-) (pause bool, resume bool, memoryUsageRatio float64) {
-	memoryUsageRatio = float64(pendingSize) / float64(maxPendingSize)
-
-	switch {
-	case paused:
-		// If the area is already paused, we need to resume it when the memory usage is less than 50%.
-		if memoryUsageRatio < 0.5 {
-			resume = true
-		}
-	default:
-		// If the area is not paused, we need to pause it when the memory usage is greater than 80% of max pending size.
-		if memoryUsageRatio >= 0.8 {
-			pause = true
-		}
-	}
-
-	return
-}
-
-func shouldPausePathV2(
-	paused bool,
-	pathPendingSize int64,
-	areaPendingSize int64,
-	maxPendingSize uint64,
-	pathCount int64,
-) (pause bool, resume bool, memoryUsageRatio float64) {
-	if pathCount == 0 {
-		log.Warn("pathCount is 0, this should not happen, adjust it to 1")
-		pathCount = 1
-	}
-
-	pathMemoryUsageRatio := float64(pathPendingSize) / float64(maxPendingSize)
-	areaMemoryUsageRatio := float64(areaPendingSize) / float64(maxPendingSize)
-
-	pauseLimit, resumeLimit := calculateThresholds(pathCount, areaMemoryUsageRatio)
-
-	if paused {
-		resume = pathMemoryUsageRatio < resumeLimit
-	} else {
-		pause = pathMemoryUsageRatio > pauseLimit
-	}
-
-	return pause, resume, pathMemoryUsageRatio
-}
-
-// shouldPauseAreaV2 determines never pause a area.
-func shouldPauseAreaV2(
-	paused bool,
-	pendingSize int64,
-	maxPendingSize uint64,
-) (pause bool, resume bool, memoryUsageRatio float64) {
-	memoryUsageRatio = float64(pendingSize) / float64(maxPendingSize)
-	return false, false, memoryUsageRatio
-}
-
-// calculateThresholds calculates dynamic pause and resume memory thresholds for path-level flow control.
-// It takes into account both the number of paths and current area memory usage to determine appropriate limits.
-// The function implements an adaptive threshold strategy:
-// - For 1-2 paths, it uses fixed thresholds
-// - For >2 paths, it:
-//  1. Applies a penalty factor that increases with path count
-//  2. Uses a tiered threshold system based on area memory usage
-//  3. Ensures minimum thresholds based on path count
-//
-// Parameters:
-//   - pathCount: number of active paths in the area
-//   - areaMemoryUsageRatio: current memory usage ratio of the area (used/max)
-//
-// Returns:
-//   - pauseLimit: memory ratio threshold to trigger path pausing
-//   - resumeLimit: memory ratio threshold to trigger path resuming
-func calculateThresholds(pathCount int64, areaMemoryUsageRatio float64) (pauseLimit, resumeLimit float64) {
-	if pathCount == 0 {
-		pathCount = 1
-	}
-
-	// Special case for path count 1 and 2
-	switch pathCount {
-	case 1:
-		return 0.8, 0.4
-	case 2:
-		return 0.5, 0.25
-	}
-
-	// calculate penalty factor based on path count
-	penaltyFactor := 1.0 - 0.5/float64(pathCount)
-
-	type threshold struct {
-		areaUsage   float64
-		pauseLimit  float64
-		resumeLimit float64
-	}
-
-	baseThresholds := []threshold{
-		//
-		{0.5, 0.2, 0.1},   // area usage <= 50%
-		{0.8, 0.1, 0.05},  // area usage <= 80%
-		{1.0, 0.05, 0.01}, // area usage <= 100%
-		// Default maxPendingSize is 1024MB, so the default lowest pause limit is 10 MB.
-		{1.2, 0.01, 0.005},  // area usage > 100%
-		{1.5, 0.005, 0.001}, // area usage > 150%, set the lowest pause limit to 1 MB.
-	}
-
-	// find applicable threshold
-	var t threshold
-	for _, th := range baseThresholds {
-		if areaMemoryUsageRatio <= th.areaUsage {
-			t = th
-			break
-		}
-	}
-
-	return t.pauseLimit / penaltyFactor, t.resumeLimit / penaltyFactor
 }
